+++ src/main/java/edu/illinois/library/cantaloupe/config/package-info.java
@@ -1,7 +1,4 @@
 /**
- * <p>Contains classes related to application configuration.</p>
- *
- * <p>This package will be needed in standalone (non-Servlet container)
- * context and should have as few dependencies as possible.</p>
+ * Contains classes related to application configuration.
  */
 package edu.illinois.library.cantaloupe.config;